+++ packages/@kagal-build-tsdoc/build.config.ts
@@ -5,6 +5,7 @@ import { newUnbuildHooks } from './src/index';
 export default defineBuildConfig({
   entries: [
     { input: 'src/index', name: 'index' },
+    { input: 'src/utils', name: 'utils' },
   ],
   declaration: true,
   sourcemap: true,